requirements.txt
moviepy==1.0.3
numpy==1.24.3
openai-whisper==20231117
torch==2.1.2
flask==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0 